# perf: reduce default follower timeout to 10ms